backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
django-cors-headers>=4.0
channels>=4.0
channels-redis>=4.2
redis>=5.0
Pillow>=10.0
psycopg2-binary>=2.9
python-dotenv>=1.0